Base code for renaming interfaces
This file is the base code for renaming interfaces on Cisco switches. The code uses a csv file which already has interfaces, descriptions, and mac-addresses listed from a or your inventory. It then loops through the switches in your switch_database file looking to see if the interfaces match the items in the csv file and if they match it will leave it alone. If it does not match it will use the data in the csv file to correct the interfaces with the correct inventoried data in your csv file as save the configs.